feat: split new flow for policy violations - Part 5
