+++ src/app/core/services/platform/v1/spender/expenses.service.ts
@@ -130,82 +130,6 @@ export class ExpensesService {
     return formattedResponse;
   }
 
-  transformSplitTo(splitTxns: Transaction[], transaction: Transaction, fileIds: string[], reportId: string) {
-    const platformSplitObject: SplitPayload = {
-      id: transaction?.id,
-      splits: this.transformSplitArray(splitTxns),
-      // Platform will throw error if category_id is null in form, therefore adding unspecified category
-      category_id: transaction?.org_category_id,
-      source: transaction?.source,
-      spent_at: transaction?.txn_dt,
-      is_reimbursable: transaction?.skip_reimbursement === null ? null : !transaction?.skip_reimbursement,
-      travel_classes: [],
-      locations: transaction?.locations,
-      foreign_currency: transaction?.orig_currency,
-      foreign_amount: transaction?.orig_amount,
-      project_id: transaction?.project_id,
-      file_ids: fileIds,
-      cost_center_id: transaction?.cost_center_id,
-      source_account_id: transaction?.source_account_id,
-      tax_amount: transaction?.tax_amount,
-      started_at: transaction?.from_dt,
-      ended_at: transaction?.to_dt,
-      merchant: transaction?.vendor,
-      purpose: transaction?.purpose,
-      is_billable: transaction?.billable,
-      custom_fields: transaction?.custom_properties,
-      claim_amount: transaction?.amount,
-    };
-
-    if (transaction?.fyle_category && transaction?.fyle_category.toLowerCase() === 'airlines') {
-      if (transaction?.flight_journey_travel_class) {
-        platformSplitObject.travel_classes.push(transaction?.flight_journey_travel_class);
-      }
-      if (transaction?.flight_return_travel_class) {
-        platformSplitObject.travel_classes.push(transaction?.flight_return_travel_class);
-      }
-    } else if (
-      transaction?.fyle_category &&
-      transaction?.fyle_category.toLowerCase() === 'bus' &&
-      transaction?.bus_travel_class
-    ) {
-      platformSplitObject.travel_classes.push(transaction?.bus_travel_class);
-    } else if (
-      transaction?.fyle_category &&
-      transaction?.fyle_category.toLowerCase() === 'train' &&
-      transaction?.train_travel_class
-    ) {
-      platformSplitObject.travel_classes.push(transaction?.train_travel_class);
-    }
-
-    if (reportId) {
-      platformSplitObject.report_id = reportId;
-    }
-
-    return platformSplitObject;
-  }
-
-  transformSplitArray(splitEtxns: Transaction[]): Splits[] {
-    const splits: Splits[] = [];
-
-    for (const splitEtxn of splitEtxns) {
-      const splitObject = {
-        spent_at: splitEtxn?.txn_dt,
-        category_id: splitEtxn?.org_category_id,
-        project_id: splitEtxn?.project_id,
-        cost_center_id: splitEtxn?.cost_center_id,
-        purpose: splitEtxn?.purpose,
-        foreign_amount: splitEtxn?.orig_amount,
-        custom_fields: splitEtxn?.custom_properties,
-        claim_amount: splitEtxn?.amount,
-      };
-
-      splits.push(splitObject);
-    }
-
-    return splits;
-  }
-
   splitExpenseCheckPolicies(params: SplitPayload) {
     return this.spenderService.post<SplitExpensePolicy>('/expenses/split/check_policies', {
       data: params,
